consult by topsis

+++ handler/model/consults.go
@@ -2,5 +2,5 @@ package model
 
 type ConsultResponse struct {
 	Name       string  `json:"name"`
-	Similarity float32 `json:"similarity"`
+	Similarity float64 `json:"similarity"`
 }
